+++ src/packs/parsing/erb/packwerk/parser.rs
@@ -3,7 +3,7 @@ use regex::Regex;
 use crate::packs::{parsing::Range, ProcessedFile, UnresolvedReference};
 use std::{fs, path::Path};
 
-use crate::packs::parsing::ruby::packwerk::extractor::process_from_contents as process_from_ruby_contents;
+use crate::packs::parsing::ruby::packwerk::parser::process_from_contents as process_from_ruby_contents;
 
 pub(crate) fn process_from_path(path: &Path) -> ProcessedFile {
     let contents = fs::read_to_string(path).unwrap_or_else(|_| {
