+++ src/packs/parsing/ruby/packwerk.rs
@@ -1,11 +1,11 @@
 pub(crate) mod constant_resolver;
-pub(crate) mod extractor;
+pub(crate) mod parser;
 
 #[cfg(test)]
 mod tests {
     use std::path::PathBuf;
 
-    use crate::packs::parsing::ruby::packwerk::extractor::process_from_contents;
+    use crate::packs::parsing::ruby::packwerk::parser::process_from_contents;
     use crate::packs::parsing::Range;
     use crate::packs::UnresolvedReference;
